2970 moving library to vendors folder
