+++ src/pageEditor/tabs/editTab/dataPanel/DataTabJsonTree.test.tsx
@@ -26,8 +26,8 @@ import { createRenderFunction } from "@/testUtils/testHelpers";
 import { DataPanelTabKey } from "./dataPanelTypes";
 import DataTabJsonTree from "./DataTabJsonTree";
 import userEvent from "@testing-library/user-event";
-import { cleanup, perf } from "@/testUtils/performanceTesting/perf";
-import { RenderCountField } from "@/testUtils/performanceTesting/perfTypes";
+import { cleanup, perf } from "@/vendors/reactPerformanceTesting/perf";
+import { RenderCountField } from "@/vendors/reactPerformanceTesting/perfTypes";
 import { act } from "@testing-library/react";
 
 jest.unmock("react-redux");
