--- src/testUtils/performanceTesting/readme.md
@@ -1,3 +1,0 @@
-# Directory
-
-This is a partial import of [react-performance-testing](https://github.com/keiya01/react-performance-testing) library that is slightly adapted for the project.
