--- src/testUtils/performanceTesting/utils/pushTask.ts
@@ -1,3 +1,0 @@
-export const pushTask = (callback: () => void) => {
-  process.nextTick(callback);
-};
