+++ src/vendors/reactPerformanceTesting/getPatchedComponent.ts
