+++ src/vendors/reactPerformanceTesting/utils/isClassComponent.ts
