+++ src/vendors/reactPerformanceTesting/utils/isForwardRefComponent.ts
