+++ src/vendors/reactPerformanceTesting/utils/isFunctionComponent.ts
