+++ src/vendors/reactPerformanceTesting/utils/isMemoComponent.ts
